Revert "Hotfix/eslint rules and errors"
