+++ .env.example
@@ -5,7 +5,3 @@ CLIENT_SECRET=
 # Spotify User credentials
 PLAYLIST_ID=
 USERNAME=
-
-# MongoDB credentials
-MONGODB_URL=
-
