+++ config.js
@@ -1,0 +1,8 @@
+require('dotenv').config();
+
+module.exports = {
+  clientId: process.env.CLIENT_ID,
+  clientSecret: process.env.CLIENT_SECRET,
+  playlistId: process.env.PLAYLIST_ID,
+  username: process.env.USERNAME,
+};
