+++ package.json
@@ -24,8 +24,6 @@
   "license": "ISC",
   "dependencies": {
     "dotenv": "^6.0.0",
-    "mongoose": "^5.2.14",
-    "mongoose-mongodb-errors": "0.0.2",
     "spotify-web-api-node": "^3.1.1"
   },
   "devDependencies": {
